Fix off by one error, fix type coercion

+++ src/Mnemonics/Mnemonics.cpp
@@ -34,7 +34,7 @@ namespace Mnemonics
     {
         Crypto::SecretKey key;
 
-        const int len = words.size();
+        const size_t len = words.size();
 
         /* Mnemonics must be 25 words long */
         if (len != 25)
@@ -83,7 +83,7 @@ namespace Mnemonics
             const uint32_t w3 = wordIndexes[i + 2];
 
             /* Word list length */
-            const uint32_t wlLen = WordList::English.size();
+            const size_t wlLen = WordList::English.size();
 
             /* no idea what this does lol */
             const uint32_t val = w1 + wlLen * (((wlLen - w1) + w2) % wlLen) + wlLen 
@@ -194,7 +194,7 @@ namespace Mnemonics
 
             /* Take it away from the beginning of the vector, giving us the
                index of the item in the vector */
-            result.push_back(std::distance(WordList::English.begin(), it));
+            result.push_back(static_cast<int>(std::distance(WordList::English.begin(), it)));
         }
 
         return result;
